refactor(core): changed kafka services interfaces
